+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Replace logrus with zerolog. (#413, @miry)
 * Log HTTP requests to API server. (#413, #421, @miry)
 * Add TimeoutHandler for the HTTP API server. (#420, @miry)
+* Set Write and Read timeouts for HTTP API server connections. (@miry)
 
 # [2.4.0] - 2022-03-07